+++ src/main/java/edu/ie3/simona/api/data/model/em/EmCommunicationMessage.java
@@ -16,8 +16,8 @@ import java.util.UUID;
  * @param msgId the message id of this message
  * @param content the actual em data that is sent
  */
-public record EmCommunicationMessage(UUID receiver, UUID sender, UUID msgId, EmData content)
-    implements EmData {
+public record EmCommunicationMessage(
+    UUID receiver, UUID sender, UUID msgId, EmMessageContent content) implements EmData {
 
   /**
    * Constructor for {@link EmCommunicationMessage} that will generate a random message id.
@@ -26,7 +26,7 @@ public record EmCommunicationMessage(UUID receiver, UUID sender, UUID msgId, EmD
    * @param sender of the data
    * @param content the actual em data that is sent
    */
-  public EmCommunicationMessage(UUID receiver, UUID sender, EmData content) {
+  public EmCommunicationMessage(UUID receiver, UUID sender, EmMessageContent content) {
     this(receiver, sender, UUID.randomUUID(), content);
   }
 }